Moving BitBuffer into .cpp; comment on cast
